rename: 타입 하위 API 관련 폴더 api -> apis로 폴더명 변경 (#173)
